# exchange-backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Exchange Rate Application",
    "description": "This is an exchange rate application where users can sign up for an account, add transactions, see their previous transactions, and get the current exchange rate calculated according to the past 3 days. Users can also check the conversion rate and amount for buying or selling USD and LBP.",
    "version": "1.0.11"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5000/"
    }
  ],
...
  "paths": {
...
    "/transaction": {
      "post": {
        "tags": [
          "Transaction"
        ],
        "summary": "Create a new transaction",
        "operationId": "createTransaction",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "usd_amount",
                  "lbp_amount",
                  "usd_to_lbp"
                ],
                "properties": {
                  "usd_amount": {
                    "type": "number",
                    "format": "float",
                    "description": "Amount in USD"
                  },
                  "lbp_amount": {
                    "type": "number",
                    "format": "float",
                    "description": "Amount in LBP"
                  },
                  "usd_to_lbp": {
                    "type": "boolean",
                    "description": "Conversion direction, USD to LBP if true"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Transaction created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          },
          "400": {
            "description": "Validation error for input data"
          },
          "403": {
            "description": "Authentication error, token invalid or expired"
          }
        },
        "security": [
          {
            "None": []
          },
          {
            "JWTToken": []
          }
        ]
      },
      "get": {
        "tags": [
          "Transaction"
        ],
        "summary": "Retrieve all transactions for the authenticated user",
        "operationId": "getTransactions",
        "responses": {
          "200": {
            "description": "A list of transactions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Transaction"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Authentication error, token invalid or expired"
          }
        },
        "security": [
          {
            "JWTToken": []
          }
        ]
      }
    },
...
    "/user": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Create new user",
        "description": "A new user can create an account for themselves.",
        "security": [],
        "operationId": "createUser",
        "requestBody": {
          "$ref": "#/components/requestBodies/UserRequestBody"
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Input empty, username not valid, or password is weak.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Weak password"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "User already exists",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "User already exists"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get user from token",
        "description": "Retrieves user information based on the provided JWT token.",
        "responses": {
          "200": {
            "description": "User retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "498": {
            "description": "Invalid or expired token."
          },
          "499": {
            "description": "Token not provided."
          }
        },
        "security": [
          {
            "JWTToken": []
          }
        ]
      }
    },
    "/user/updateUserPreferences": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Update user preferences. READ DESCRIPTION IMPORTANT!!!!!!!!!!!!!!",
        "description": "Updates user preferences for receiving notifications based on exchange rate changes. NOTE THAT IF YOU CHANGE ONE OF THE BOOLEAN FLAGS TO TRUE WITHOUT SETTING ALL THE PARAMETERS OF IT, PREVIOUS PREFERENCES WILL BE OVERRIDEN. WHAT I MEAN IS: ASSUME A USEER SET an upper_bound_usd_to_lbp and saves. Then the user wants to send lower_bound_usd_to_lbp. IF YOU SEND THE JSON WITH  upper_bound_usd_to_lbp SET TO NULL, IT WILL BE DELETED AND USER WON't BE NOTIFIED WHEN IT IS FULFILLED. SIMILARLY FOR RATES. AS FOR EMAILS AND SMS IF YOU SET update_email_preference = true and send_email = true AN EMAIL NEEDS TO BE SENT TO BACKEND REGARDLESS IF THERE IS ONE ALREADY IN THE DATABASE. SIMILARLY FOR PHONE NUMBER. SO, YOU CAN FETCH PREVIOUS PREFERENCES TO SEND THE CORRECT UPDATES. \n  ",
        "requestBody": {
          "$ref": "#/components/requestBodies/UserPreferences"
        },
        "responses": {
          "200": {
            "description": "User preferences updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPreferences"
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid data provided."
          },
          "498": {
            "description": "Invalid token."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "JWTToken": []
          }
        ]
      }
    },
    "/user/getUserPreferencesFromId": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get user preferences",
        "description": "Retrieves the user preferences for notifications.",
        "responses": {
          "200": {
            "description": "User preferences retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPreferences"
                }
              }
            }
          },
          "498": {
            "description": "Invalid token."
          },
          "499": {
            "description": "Token not provided."
          }
        },
        "security": [
          {
            "JWTToken": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "UserPreferences": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "user_id": {
            "type": "integer",
            "example": 4
          },
          "upper_bound_usd_to_lbp": {
            "type": "number",
            "example": 90000.67
          },
          "lower_bound_usd_to_lbp": {
            "type": "number",
            "example": 85000
          },
          "upper_bound_lbp_to_usd": {
            "type": "number",
            "example": 85000
          },
          "lower_bound_lbp_to_usd": {
            "type": "number",
            "example": 85000
          },
          "email": {
            "type": "string",
            "example": "hello@gmail.com"
          },
          "send_email": {
            "type": "boolean"
          },
          "phone_number": {
            "type": "string",
            "example": "+96101111111"
          },
          "send_sms": {
            "type": "boolean",
            "example": false
          }
        }
      },
      "Transaction": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "The transaction identifier"
          },
          "usd_amount": {
            "type": "number",
            "format": "float",
            "description": "Amount in USD"
          },
          "lbp_amount": {
            "type": "number",
            "format": "float",
            "description": "Amount in LBP"
          },
          "usd_to_lbp": {
            "type": "boolean",
            "description": "Conversion direction, USD to LBP if true"
          },
          "added_date": {
            "type": "string",
            "format": "date-time",
            "description": "The date and time when the transaction was added"
          },
          "user_id": {
            "type": "integer",
            "description": "The identifier of the user who made the transaction"
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "user_name": {
            "type": "string",
            "example": "omar1212"
          }
        }
      },
...
    },
    "requestBodies": {
      "UserPreferences": {
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "update_bounds": {
                  "type": "boolean",
                  "example": true
                },
                "upper_bound_usd_to_lbp": {
                  "type": "number",
                  "example": 90000.67
                },
                "lower_bound_usd_to_lbp": {
                  "type": "number",
                  "example": 85000
                },
                "upper_bound_lbp_to_usd": {
                  "type": "number",
                  "example": 85000
                },
                "lower_bound_lbp_to_usd": {
                  "type": "number",
                  "example": 85000
                },
                "update_email_preference": {
                  "type": "boolean",
                  "example": true
                },
                "email": {
                  "type": "string",
                  "example": "hello@gmail.com"
                },
                "send_email": {
                  "type": "boolean"
                },
                "update_sms_preference": {
                  "type": "boolean",
                  "example": true
                },
                "phone_number": {
                  "type": "string",
                  "example": "+96101111111"
                },
                "send_sms": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      },
...
      "UserRequestBody": {
        "description": "User data for registration",
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "user_name": {
                  "type": "string",
                  "example": "omar1212"
                },
                "password": {
                  "type": "string",
                  "format": "password",
                  "example": "Pass_1212"
                }
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "JWTToken": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "JSON Web Token (JWT) allow users to link transactions to themselves.\nAdd the JWT token in the format `Bearer <token>`.\n"
      },
      "None": {
        "type": "http",
        "scheme": "none",
        "description": "Authentication is not needed always.\n"
      }
    }
  }
}